requirements.txt
matplotlib==3.4.3
numpy==1.20.3
opencv_python==4.5.4.60
pandas==1.3.4
Pillow==9.1.0
pytorch_tabnet==3.1.1
scikit_learn==1.1.0
scipy==1.7.1
timm==0.5.4
torch==1.10.0+cu113
tqdm==4.62.3
